// src/api-services/models/set-user-met-meet-record-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 恒瑞代谢公众号
 * 恒瑞代谢相关公众号接口
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface SetUserMetMeetRecordInput
 */
export interface SetUserMetMeetRecordInput {
    /**
     * 会议ID
     * @type {number}
     * @memberof SetUserMetMeetRecordInput
     */
    meetId: number;
    /**
     * 观看时长
     * @type {number}
     * @memberof SetUserMetMeetRecordInput
     */
    duration: number;
}